+++ CHANGELOG.md
@@ -2,7 +2,22 @@
 
 All notable changes to `nova-logs-tool` will be documented in this file
 
-## 1.0.0 - 2022-04-13
+## 2.0.0
+
+- Now our package is based on [laravel-log-viewer](https://github.com/stepanenko3/laravel-log-viewer)
+- Drop LogsService class
+- Dropping our configs and using them from [laravel-log-viewer](https://github.com/stepanenko3/laravel-log-viewer)
+- Added log levels
+- Support regex in search
+- Show log file size
+- Logs and files sorting
+- Custom pagination
+- Show line index
+- Show environment
+- Show loading time and memory
+- Other fixes and updates
+
+## 1.0.0
 
 - Nova 4 support
 - initial release